% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualize.R
\name{annualize}
\alias{annualize}
\title{Function to convert daily water budget data to an annualzed time series.}
\usage{
annualize(inputData, method = mean)
}
\arguments{
\item{inputData}{A daily data.frame containing a Date, format
"yyyy-mm-dd" 'date' and numerical 'data' collumn.}

\item{method}{A function to use in grouping. Defaults to mean, can be sum or other statistic.}
}
\value{
An annual time series data frame with a year and data collumn.
}
\description{
Basic function that takes a date/data data frome and returns an annualized time series.
}
\examples{
data<-getSWECSVBlock(system.file('extdata','SWECSVBlock_daymet_example.xml',package='NWCEd'))
annual<-annualize(data)

}

