% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNWCData.R
\name{getNWCData}
\alias{getNWCData}
\title{Get data from an NWC source.}
\usage{
getNWCData(huc)
}
\arguments{
\item{huc}{The watershed of interest.}
}
\value{
The data.
}
\description{
This function builds a request and returns the the data in question.
}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}

