% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdata.R
\name{plotdata}
\alias{plotdata}
\title{Plot data from an NWC source.}
\usage{
plotdata(Downloaded_Data)
}
\arguments{
\item{Downloaded_Data}{The downloaded data to be plotted.}
}
\value{
The plot.
}
\description{
This function builds a request and returns the the data in question.
}
\examples{
data<-getNWCData(huc="031601030306")
}
\author{
David Blodgett \email{dblodgett@usgs.gov}
}

