% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualize.R
\name{annualize}
\alias{annualize}
\title{Function to convert daily water budget data to an annualzed time series.}
\usage{
annualize(inputData)
}
\arguments{
\item{inputData}{A daily data.frame containing a Date, format
"yyyy-mm-dd" 'date' and numerical 'data' collumn.}
}
\value{
An annual time series data frame with a year and data collumn.
}
\description{
Basic function that takes a date/data data frome and returns an annualized time series.
}
\examples{
url <- "http://cida.usgs.gov/nwc/thredds/sos/watersmart/HUC12_data/HUC12_daymet.nc?request=GetObservation&service=SOS&version=1.0.0&observedProperty=prcp&offering=031401020800"
data<-getSWECSVBlock(url)
annual<-annualize(data)

}

