% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSWECSVBlock.R
\name{getSWECSVBlock}
\alias{getSWECSVBlock}
\title{Function to return NWC Intercomparison portal modeled data for a given site}
\usage{
getSWECSVBlock(input)
}
\arguments{
\item{input}{url for SOS service for desired site data}
}
\value{
data frame containing desired time series
}
\description{
This function accepts a url and returns a data frame of data for that SOS request
}
\examples{
getSWECSVBlock(system.file('extdata','SWECSVBlock_daymet_example.xml',package='NWCEd'))
}

