% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lp3.R
\name{Lp3}
\alias{Lp3}
\title{The Lp3 function applies the Log-Pearson Type III model to hydrologic datasets from the National Water Census Data Portal to generate frequency distribution curves.}
\usage{
Lp3(inputData, datatype)
}
\arguments{
\item{inputData}{\code{data.frame} stored in user-named variable, returned by getNWCData}

\item{datatype}{\code{character} choose from "streamflow" and "prcp"}
}
\value{
The return variable "plotthis" is a plot of the Log-Pearson Type III model applied to the user-selected USGS hydrologic dataset.
}
\description{
The purpose of the Lp3 function is to apply the Log-Pearson Type III model to streamflow and precipitation datasets from the National Water Census Data Portal to obtain their respective frequency distribution curves.
The desired datasets are to be previously downloaded from the Portal using the 'getNWCData' function and saved in a user-named variable. The user inputs two arguments into the Lp3 function: inputData (the user-named variable) and the desired dataset ("streamflow" or "prcp").
The function singles out the desired dataset and and converts units from metric to english. The dataset is organized into water years. Next, the Log Pearson Type III model is applied to the data. The function calls in Frequency Factors from a csv.
Freqency factors were taken from http://streamflow.engr.oregonstate.edu/analysis/floodfreq/skew.htm.
}
\examples{
\dontrun{
variable_name<-getNWCData(huc = "160202030505", local = FALSE)
Lp3(variable_name, "prcp")
}
}

